Changed variable definitions, random spelling fixes
Use 16bit shorts instead of 32bit ints everywhere. HUGE waste of space for absolutely nothing
I'm also hoping that declaring some arrays as constant will store it *only* in program Flash. Not sure if that's the case yet.
Made almost all chars into unsigned chars. Not sure what the real repercussions are for strings, but I feel better about this. It's what I want/expect them to be.

+++ code/settings.h
@@ -37,7 +37,7 @@ typedef enum DELTA
 	NONE
 } DELTA;
 
-void playNote(unsigned int);
+void playNote(unsigned short int);
 void configure();
 
 #endif
